requirements.txt
click==8.2.1
dask==2025.7.0
fiona==1.10.1
geopandas==1.1.1
h3==4.3.0
libpysal==4.13.0
matplotlib==3.10.3
mgwr==2.2.1
minisom==2.3.5
numpy==2.3.1
osgeo==0.0.1
pandas==2.3.1
psutil==7.0.0
psycopg2==2.9.10
pyproj==3.7.1
pytest==8.4.1
PyYAML==6.0.2
rasterio==1.4.3
rioxarray==0.19.0
scikit_learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
Shapely==2.1.1
spopt==0.7.0
tqdm==4.67.1
xarray==2025.7.1
